Move filesystem types to corresponding namespace
