+++ src/RazorLight/Templating/LoadedTemplateSource.cs
@@ -3,25 +3,34 @@ using System.IO;
 
 namespace RazorLight.Templating
 {
-    public class StringTemplateSource : ITemplateSource
+    public class LoadedTemplateSource : ITemplateSource
     {
 	    public string Content { get; private set; }
 	    public string FilePath { get; private set; }
 	    public string TemplateKey { get; }
 
-	    public StringTemplateSource(string content)
+	    public LoadedTemplateSource(string key) : this(key, GetRandomString())
 	    {
-		    if (string.IsNullOrEmpty(content))
+	    }
+
+	    public LoadedTemplateSource(string key, string content)
+	    {
+			if (string.IsNullOrEmpty(key))
+			{
+				throw new ArgumentNullException(nameof(key));
+			}
+
+			if (string.IsNullOrEmpty(content))
 		    {
 			    throw new ArgumentNullException(nameof(content));
 		    }
 
 		    this.Content = content;
-		    this.TemplateKey = GetRandomString();
-		    this.FilePath = null;
+		    this.TemplateKey = key;
+			this.FilePath = null;
 	    }
 
-	    private string GetRandomString()
+	    private static string GetRandomString()
 	    {
 		    return Path.GetFileName(Path.GetRandomFileName());
 	    }
